Correct routes for events booking
