+++ src/Payutc/OnyxBundle/Entity/Price.php
@@ -158,14 +158,37 @@ class Price extends BaseEntity implements Serializable
         return $this;
     }
 
+    /**
+     * Magic __toString function
+     * Returns the title of the price
+     *
+     * @return string
+     */
     public function __toString()
     {
         return $this->getTitle();
     }
 
+    /**
+     * toString function
+     * Returns the title of the price
+     *
+     * @return string
+     */
     public function toString()
     {
         return $this->getTitle();
+    }
+
+    /**
+     * Get title and price
+     * Form helper
+     *
+     * @return string
+     */
+    public function getTitleAndPrice()
+    {
+        return $this->getTitle() . ' - ' . $this->getPrice() . '€';
     }
 
     /**
@@ -191,7 +214,7 @@ class Price extends BaseEntity implements Serializable
     /**
      * Entity has been removed or not ?
      *
-     * @return boolean 
+     * @return boolean
      */
     public function isDeleted()
     {
@@ -201,7 +224,7 @@ class Price extends BaseEntity implements Serializable
     /**
      * Get id
      *
-     * @return integer 
+     * @return integer
      */
     public function getId()
     {
@@ -217,14 +240,14 @@ class Price extends BaseEntity implements Serializable
     public function setCreatedAt($createdAt)
     {
         $this->createdAt = $createdAt;
-    
+
         return $this;
     }
 
     /**
      * Get createdAt
      *
-     * @return \DateTime 
+     * @return \DateTime
      */
     public function getCreatedAt()
     {
@@ -240,14 +263,14 @@ class Price extends BaseEntity implements Serializable
     public function setUpdatedAt($updatedAt)
     {
         $this->updatedAt = $updatedAt;
-    
+
         return $this;
     }
 
     /**
      * Get updatedAt
      *
-     * @return \DateTime 
+     * @return \DateTime
      */
     public function getUpdatedAt()
     {
@@ -263,14 +286,14 @@ class Price extends BaseEntity implements Serializable
     public function setRemovedAt($removedAt)
     {
         $this->removedAt = $removedAt;
-    
+
         return $this;
     }
 
     /**
      * Get removedAt
      *
-     * @return \DateTime 
+     * @return \DateTime
      */
     public function getRemovedAt()
     {
@@ -286,14 +309,14 @@ class Price extends BaseEntity implements Serializable
     public function setIsHidden($isHidden)
     {
         $this->isHidden = $isHidden;
-    
+
         return $this;
     }
 
     /**
      * Get isHidden
      *
-     * @return boolean 
+     * @return boolean
      */
     public function getIsHidden()
     {
@@ -309,14 +332,14 @@ class Price extends BaseEntity implements Serializable
     public function setStartAt($startAt)
     {
         $this->startAt = $startAt;
-    
+
         return $this;
     }
 
     /**
      * Get startAt
      *
-     * @return \DateTime 
+     * @return \DateTime
      */
     public function getStartAt()
     {
@@ -332,14 +355,14 @@ class Price extends BaseEntity implements Serializable
     public function setEndAt($endAt)
     {
         $this->endAt = $endAt;
-    
+
         return $this;
     }
 
     /**
      * Get endAt
      *
-     * @return \DateTime 
+     * @return \DateTime
      */
     public function getEndAt()
     {
@@ -355,14 +378,14 @@ class Price extends BaseEntity implements Serializable
     public function setTitle($title)
     {
         $this->title = $title;
-    
+
         return $this;
     }
 
     /**
      * Get title
      *
-     * @return string 
+     * @return string
      */
     public function getTitle()
     {
@@ -378,14 +401,14 @@ class Price extends BaseEntity implements Serializable
     public function setDescription($description)
     {
         $this->description = $description;
-    
+
         return $this;
     }
 
     /**
      * Get description
      *
-     * @return string 
+     * @return string
      */
     public function getDescription()
     {
@@ -401,14 +424,14 @@ class Price extends BaseEntity implements Serializable
     public function setPrice($price)
     {
         $this->price = $price;
-    
+
         return $this;
     }
 
     /**
      * Get price
      *
-     * @return integer 
+     * @return integer
      */
     public function getPrice()
     {
@@ -424,14 +447,14 @@ class Price extends BaseEntity implements Serializable
     public function setIsNominativ($isNominativ)
     {
         $this->isNominativ = $isNominativ;
-    
+
         return $this;
     }
 
     /**
      * Get isNominativ
      *
-     * @return boolean 
+     * @return boolean
      */
     public function getIsNominativ()
     {
@@ -447,14 +470,14 @@ class Price extends BaseEntity implements Serializable
     public function setCapacity($capacity)
     {
         $this->capacity = $capacity;
-    
+
         return $this;
     }
 
     /**
      * Get capacity
      *
-     * @return integer 
+     * @return integer
      */
     public function getCapacity()
     {
@@ -470,14 +493,14 @@ class Price extends BaseEntity implements Serializable
     public function setMaxPlacesForUser($maxPlacesForUser)
     {
         $this->maxPlacesForUser = $maxPlacesForUser;
-    
+
         return $this;
     }
 
     /**
      * Get maxPlacesForUser
      *
-     * @return integer 
+     * @return integer
      */
     public function getMaxPlacesForUser()
     {
@@ -493,14 +516,14 @@ class Price extends BaseEntity implements Serializable
     public function setIsUniqueForUser($isUniqueForUser)
     {
         $this->isUniqueForUser = $isUniqueForUser;
-    
+
         return $this;
     }
 
     /**
      * Get isUniqueForUser
      *
-     * @return boolean 
+     * @return boolean
      */
     public function getIsUniqueForUser()
     {
@@ -516,14 +539,14 @@ class Price extends BaseEntity implements Serializable
     public function setPayutcId($payutcId)
     {
         $this->payutcId = $payutcId;
-    
+
         return $this;
     }
 
     /**
      * Get payutcId
      *
-     * @return integer 
+     * @return integer
      */
     public function getPayutcId()
     {
@@ -539,14 +562,14 @@ class Price extends BaseEntity implements Serializable
     public function setIsNominative($isNominative)
     {
         $this->isNominative = $isNominative;
-    
+
         return $this;
     }
 
     /**
      * Get isNominative
      *
-     * @return boolean 
+     * @return boolean
      */
     public function getIsNominative()
     {
@@ -562,7 +585,7 @@ class Price extends BaseEntity implements Serializable
     public function addUserGroup(UserGroup $userGroup)
     {
         $this->userGroups[] = $userGroup;
-    
+
         return $this;
     }
 
@@ -595,14 +618,14 @@ class Price extends BaseEntity implements Serializable
     public function setEvent(Event $event = null)
     {
         $this->event = $event;
-    
+
         return $this;
     }
 
     /**
      * Get event
      *
-     * @return Event 
+     * @return Event
      */
     public function getEvent()
     {
