+++ src/Payutc/OnyxBundle/Resources/config/routing/entities/events.yml
@@ -4,3 +4,17 @@ pay_utc_onyx_event_page:
         _controller: PayutcOnyxBundle:Entities/Event:detail
     requirements:
         _method: get
+
+pay_utc_onyx_event_ticket_book_page:
+    pattern:  /{id}/tickets/book
+    defaults:
+        _controller: PayutcOnyxBundle:Entities/Event:book
+    requirements:
+        _method: get
+
+pay_utc_onyx_event_ticket_book_check_page:
+    pattern:  /{id}/tickets/book-check
+    defaults:
+        _controller: PayutcOnyxBundle:Entities/Event:bookCheck
+    requirements:
+        _method: post
