+++ src/Payutc/OnyxBundle/Resources/views/Entities/Events/detail.html.twig
@@ -72,7 +72,7 @@
 						{% endif %}
 					{% endfor %}
 					<div class='link-wrapper'>
-						<a href='{{ path('pay_utc_onyx_ticket_book_page', { 'eventId': event.id }) }}' class='btn btn-info'>Réserver</a>
+						<a href='{{ path('pay_utc_onyx_event_ticket_book_page', { 'id': event.id }) }}' class='btn btn-info'>Réserver</a>
 					</div>
 				</div>
 			</div>
