+++ src/Payutc/OnyxBundle/Resources/views/Entities/Tickets/detail.mail.html.twig
@@ -1,0 +1,12 @@
+{% extends 'PayutcOnyxBundle::mail.html.twig' %}
+
+{% block page_title %}Billeterie UTC - Votre place : {{ ticket.price.event.title }}{% endblock %}
+
+{% block app_content %}
+    <div class='jumbotron'>
+        <h1>Bonjour {{ ticket.firstname|title }} {{ ticket.lastname|title }}</h1>
+    </div>
+    <p>Veuillez trouver votre place en pièce jointe à ce mail.</p>
+    <p>Vous pouvez également retrouver le détail de l'évènement en ligne : <a href='{{ url('pay_utc_onyx_event_page', { 'id': ticket.price.event.id }) }}' target='_blank'>{{ ticket.price.event.title }}</a>.</p>
+    <p>En espérant vous voir très vite !</p>
+{% endblock %}
