+++ src/Payutc/OnyxBundle/Resources/views/Entities/Tickets/detail.pdf.html.twig
@@ -2,7 +2,7 @@
 
 {% block app_content %}
     <div class='page-title'>
-        <h1>Votre billet : {{ event.title }}</h1>
+        <h1>Votre billet : {{ ticket.price.event.title }}</h1>
     </div>
     <p class='lead'>Bonjour {{ ticket.buyer.firstname }} {{ ticket.buyer.name }}</p>
     <p>
